blogservices/recommendation/requirements.txt
python-dotenv
pika
pymongo
nltk
requests
fastapi[standard]
aiormq==6.8.1
python-jose[cryptography]
httpx==0.28.1